pyproject.toml
[project]
name = "miniwow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<=3.12"
dependencies = [
    "peewee>=3.17.0",
    "pytest>=9.0.2",
    "imagehash>=4.3.2",
    "pytest-timeout>=2.4.0",
    "wrapt-timeout-decorator>=1.5.1",
    "setuptools>=80.9.0",
    "coloredlogs>=15.0.1",
    "airtest>=1.4.3",
    "transitions>=0.9.0",
    "streamlit>=1.52.2",
    "pandas>=2.2.0",
    "streamlit-autorefresh>=1.0.1",
    "requests>=2.32.5",
    "typer>=0.12.3",
    "opencv-python-headless>=4.11.0.86",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "vibe-ocr>=0.1.1",
    "vibe-logger>=0.1.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 100

[tool.uv]
# 声明你需要支持的平台集合：macOS arm64 + Windows x64（可选再加 Windows arm64）
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'win32'",
    # 如果你也要支持 Windows on ARM，可打开这一行
    # "sys_platform == 'win32' and platform_machine == 'ARM64'",
]

# 关键：把 pywin32 覆盖成“仅 Windows 安装”
# 这样就算有上游把 pywin32 误写成跨平台依赖，也会被你这里纠正回去
override-dependencies = ["pywin32==311; sys_platform == 'win32'"]
